:heavy_minus_sign: Removed boost requirements from the Lorem module
used stdlib functionalities to supplant boost::algorithm

+++ include/faker-cxx/Lorem.h
@@ -8,6 +8,6 @@ class Lorem
 {
 public:
     static std::string word();
-    static std::string words(int numberOfWords = 3);
+    static std::string words(size_t numberOfWords = 3);
 };
 }
